group_metrics
